Now shows the confidance

+++ source/c++/src/boundingBox.cpp
@@ -6,7 +6,7 @@ vector<BoundingBox> generateBoundingBoxes(uint16_t* points, uint16_t* types, flo
 	for (uint16_t i = 0; i < size; i++){
         uint16_t* currentPoint = points + i*core::_elementsInPoint;
         uint16_t currentType = types[i];
-        uint16_t currentConfidance = confidances[i];
+        float32 currentConfidance = confidances[i];
         boundingBoxes.emplace_back(Rect(currentPoint), currentType, currentConfidance);
 	}
